+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 =================
 
 ## 0.18.71
+### is.codion.common.reactive
+- State.Group.fallback() added, related refactoring.
 ### is.codion.common.utilities
 - User is now sealed.
 ### is.codion.common.rmi